+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- 🐛(resource-server) allow `aud` & `iss` JWE headers #24
+
 ## [0.0.10] - 2025-06-18
 
 ### Fixed